+++ src/onegov.activity/onegov/activity/tests/test_models.py
@@ -277,6 +277,75 @@ def test_activity_date_ranges(session, owner, collections):
     assert a.query().count() == 1
 
 
+def test_activity_weekdays(session, owner, collections):
+    sport = collections.activities.add("Sport", username=owner.username)
+    police = collections.activities.add("Police", username=owner.username)
+
+    period = collections.periods.add(
+        title="Spring 2017",
+        prebooking=(datetime(2017, 2, 1), datetime(2017, 2, 28)),
+        execution=(datetime(2017, 5, 8), datetime(2017, 5, 21)),
+        active=True
+    )
+
+    # Monday
+    collections.occasions.add(
+        start=datetime(2017, 5, 8, 10),
+        end=datetime(2017, 5, 8, 12),
+        timezone="Europe/Zurich",
+        age=(6, 9),
+        spots=(2, 10),
+        activity=sport,
+        period=period
+    )
+
+    # Tuesday
+    collections.occasions.add(
+        start=datetime(2017, 5, 9, 10),
+        end=datetime(2017, 5, 9, 12),
+        timezone="Europe/Zurich",
+        age=(6, 9),
+        spots=(2, 10),
+        activity=sport,
+        period=period
+    )
+
+    # Tuesday - Wednesdy
+    collections.occasions.add(
+        start=datetime(2017, 5, 9, 18),
+        end=datetime(2017, 5, 10, 18),
+        timezone="Europe/Zurich",
+        age=(6, 9),
+        spots=(2, 10),
+        activity=police,
+        period=period
+    )
+
+    transaction.commit()
+
+    a = collections.activities
+
+    # monday
+    a = a.for_filter(weekday=0)
+    assert a.query().count() == 1
+
+    # monday, tuesday
+    a = a.for_filter(weekday=1)
+    assert a.query().count() == 2
+
+    # tuesday (negates monday)
+    a = a.for_filter(weekday=0)
+    assert a.query().count() == 2
+
+    # tuesday, wednesday
+    a = a.for_filter(weekday=2)
+    assert a.query().count() == 2
+
+    # wednesday (negates tuesday)
+    a = a.for_filter(weekday=1)
+    assert a.query().count() == 1
+
+
 def test_profiles(session, owner):
 
     activities = ActivityCollection(session)